Update task service and tests: add @types/cucumber, refactor ObjectId handling, and streamline step definitions
