+++ package.json
@@ -29,6 +29,7 @@
   "devDependencies": {
     "@cucumber/cucumber": "^11.1.1",
     "@types/cors": "^2.8.17",
+    "@types/cucumber": "^6.0.1",
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.9.0",
